+++ docs/changelog.rst
@@ -4,6 +4,11 @@ Change log
 Unreleased_
 -----------
 
+2.3.0_ |--| 2022-10-25
+----------------------
+
+Be thread-safe where possible.
+
 Added
 ~~~~~
 
@@ -179,7 +184,8 @@ Added
 
 .. |--| unicode:: U+2013 .. EN DASH
 
-.. _Unreleased: https://github.com/nomis/mcu-uuid-log/compare/2.2.0...HEAD
+.. _Unreleased: https://github.com/nomis/mcu-uuid-log/compare/2.3.0...HEAD
+.. _2.3.0: https://github.com/nomis/mcu-uuid-log/compare/2.2.0...2.3.0
 .. _2.2.0: https://github.com/nomis/mcu-uuid-log/compare/2.1.4...2.2.0
 .. _2.1.4: https://github.com/nomis/mcu-uuid-log/compare/2.1.3...2.1.4
 .. _2.1.3: https://github.com/nomis/mcu-uuid-log/compare/2.1.2...2.1.3